+++ Example/ViewController.swift
@@ -144,6 +144,7 @@ class ViewController: UIViewController {
         clearMap.isHidden = true
         longPressHintView.isHidden = false
         
+        // TODO: adjust camera to location after cancle
         navigationMapView?.unhighlightBuildings()
         navigationMapView?.removeRoutes()
         navigationMapView?.removeRouteDurations()
@@ -258,7 +259,7 @@ class ViewController: UIViewController {
         // Control floating buttons position in a navigation view.
         navigationViewController.floatingButtonsPosition = .topTrailing
         
-        presentAndRemoveMapview(navigationViewController)
+        presentAndRemoveMapview(navigationViewController, completion: beginCarPlayNavigation)
     }
     
     func startCustomNavigation() {
@@ -531,6 +532,11 @@ extension ViewController: RouteVoiceControllerDelegate {
 extension ViewController: NavigationViewControllerDelegate {
 
     func navigationViewController(_ navigationViewController: NavigationViewController, didArriveAt waypoint: Waypoint) -> Bool {
+        if #available(iOS 12.0, *),
+           let delegate = UIApplication.shared.delegate as? AppDelegate,
+           let carPlayNavigationViewController = delegate.carPlayManager.currentNavigator {
+            return carPlayNavigationViewController.navigationService(navigationViewController.navigationService, didArriveAt: waypoint)
+        }
         return true
     }